Fixup: move query expansion from recrawling/search-logging to query generation
